Refactor asset loading
