+++ desktop/src/main/vmixHelpers.ts
@@ -1,8 +1,9 @@
 import invariant from "../common/invariant";
 import { getVMixConnection } from "./vmix";
-import { z } from "zod";
-import { CompleteAssetSchema } from "@bowser/prisma/utilityTypes";
-import { InputType } from "./vmixTypes";
+import { InputType, ListInput } from "./vmixTypes";
+import type { Asset, Media } from "@bowser/prisma/client";
+import { getAssetsSettings, getLocalMediaSettings } from "./settings";
+import { VMIX_NAMES } from "../common/constants";
 
 export async function reconcileList(listName: string, elements: string[]) {
   const conn = getVMixConnection();
@@ -26,7 +27,7 @@ export async function reconcileList(listName: string, elements: string[]) {
 }
 
 export function getInputTypeForAsset(
-  asset: z.infer<typeof CompleteAssetSchema>,
+  asset: Asset & { media: Media | null },
 ): InputType {
   switch (asset.type) {
     case "Still":
@@ -45,5 +46,56 @@ export function getInputTypeForAsset(
     }
     default:
       invariant(false, `Unknown asset type ${asset.type}`);
+  }
+}
+
+export async function loadAssets(assets: (Asset & { media: Media | null })[]) {
+  const localMedia = await getLocalMediaSettings();
+  const settings = await getAssetsSettings();
+  const vmix = getVMixConnection();
+  invariant(vmix, "No vMix connection");
+  const state = await vmix.getFullState();
+
+  for (const asset of assets) {
+    if (!asset.media || asset.media.state !== "Ready") {
+      throw new Error("Not all assets have remote media");
+    }
+    const local = localMedia.find((x) => x.mediaID === asset.media!.id);
+    if (!local) {
+      throw new Error("No local media for asset " + asset.id);
+    }
+
+    const loadType = settings.loadTypes[asset.type];
+    invariant(
+      loadType,
+      "no load type configured for " + asset.type + " asset " + asset.id,
+    );
+    if (loadType === "direct") {
+      const present = state.inputs.find((x) => x.title === asset.media!.name);
+      if (!present) {
+        await vmix.addInput(getInputTypeForAsset(asset), local.path);
+      }
+    } else if (loadType === "list") {
+      let present;
+      const list = state.inputs.find(
+        (x) => x.shortTitle === VMIX_NAMES.ASSET_LIST[asset.type],
+      );
+      let listKey;
+      if (!list) {
+        present = false;
+        listKey = await vmix.addInput("VideoList", "");
+        await vmix.renameInput(listKey, VMIX_NAMES.ASSET_LIST[asset.type]);
+      } else {
+        listKey = list.key;
+        present = (list as ListInput).items.some(
+          (x) => x.source === local.path,
+        );
+      }
+      if (!present) {
+        await vmix.addInputToList(listKey, local.path);
+      }
+    } else {
+      invariant(false, "Invalid load type " + loadType);
+    }
   }
 }
